Split command methods off to their own files
